Fix-it C.3 (iter-04): merge PoolMetrics into single shipped source file
Resolves the locked-decision conflict flagged in the review after iter-03.
PROJECT_CONTEXT.md line 78 mandates source-shipping as ONE .cs file; the
content-only NuGet pack only includes Helios.Concurrency.DedicatedThreadPool.cs.
PoolMetrics.cs was a second core file that would have caused a missing-type
compile error for source-ship consumers the moment Phase 3 wired it in.

Changes:
- Add 'using System.Diagnostics.Metrics;' to the shipped file
- Append internal static class PoolMetrics (Meter + 3 instruments) to the
  shipped file, inside the Helios.Concurrency namespace
- Delete PoolMetrics.cs

Verification: build 0/0 warnings; 7/7 xUnit pass (1 pre-existing skip) on
net10.0; PoolMetrics_PublishesExpectedInstruments passes.
